+++ internal/log.go
@@ -6,3 +6,5 @@ import (
 )
 
 var Logger = log.New(os.Stderr, "pg: ", log.LstdFlags|log.Lshortfile)
+
+var Deprecated = log.New(os.Stderr, "DEPRECATED: pg: ", log.LstdFlags)